+++ src/renderer_app.cpp
@@ -16,15 +16,15 @@ int main(int argc, char* argv[])
     unique_ptr<Color> pixelData{ new Color[WIDTH * HEIGHT] };
     Color* pixels = pixelData.get();
 
-    for (float t = 0.1; t > 0; t -= 0.1f) {
+    for (float t = 0.1f; t > 0; t -= 0.1f) {
         // Fill the pixel data
         render(pixels, t);
         PixelBuffer img(WIDTH, HEIGHT);
         for (int y = 0; y < HEIGHT; y++) {
             for (int x = 0; x < WIDTH; x++) {
-                int R = pixels[(y*WIDTH + x)].r * 255.999f;
-                int G = pixels[(y*WIDTH + x)].g * 255.999f;
-                int B = pixels[(y*WIDTH + x)].b * 255.999f;
+                int R = int(pixels[(y*WIDTH + x)].r * 255.999f);
+                int G = int(pixels[(y*WIDTH + x)].g * 255.999f);
+                int B = int(pixels[(y*WIDTH + x)].b * 255.999f);
                 DWORD col = (R << 16) + (G << 8) + B;
                 window.setBufferPixel(x, y, col);
             }
